Add an optional attribute to handle updates during atlas downloads (#126)
* Add an optional attribute to handle updates during atlas downloads

Atlas download progress is currently handled within a Rich progress indicator, but external libraries do not have access to these updates. Add an optional attribute to `BrainGlobeAtlas` to reference a handler that is updated during atlas downloads. This handler is a function that takes the currently competed and total required bytes to download the given atlas.

* Fix formatting errors picked up by black package

- Remove whitespace on empty line
- Format line width, indentation, and closing parenthesis placement